requirements.txt
asgiref==3.4.1
Django==3.2.6
django-camera-imagefield==0.2
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
djangorestframework==3.13.1
install==1.3.4
Pillow==8.3.2
psycopg2==2.9.1
python-datauri==1.1.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.2
whitenoise==6.0.0
